move ordered transaction schema load to generate_schema_documentation.py where it's needed
