+++ docs/server/generate_schema_documentation.py
@@ -1,10 +1,11 @@
 """ Script to render transaction schema into .rst document """
 
-import datetime
-
+from collections import OrderedDict
 import os.path
 
-from bigchaindb.common.schema import TX_SCHEMA_ORDERED as TX
+import yaml
+
+from bigchaindb.common.schema import TX_SCHEMA_YAML
 
 
 TPL_PROP = """\
@@ -18,7 +19,7 @@ TPL_PROP = """\
 
 
 TPL_DOC = """\
-..  This file was auto generated by %(file)s on %(date)s
+..  This file was auto generated by %(file)s
 
 ==================
 Transaction Schema
@@ -86,6 +87,23 @@ Metadata
 """
 
 
+def ordered_load_yaml(stream):
+    """ Custom YAML loader to preserve key order """
+    class OrderedLoader(yaml.SafeLoader):
+        pass
+
+    def construct_mapping(loader, node):
+        loader.flatten_mapping(node)
+        return OrderedDict(loader.construct_pairs(node))
+    OrderedLoader.add_constructor(
+        yaml.resolver.BaseResolver.DEFAULT_MAPPING_TAG,
+        construct_mapping)
+    return yaml.load(stream, OrderedLoader)
+
+
+TX_SCHEMA = ordered_load_yaml(TX_SCHEMA_YAML)
+
+
 DEFINITION_BASE_PATH = '#/definitions/'
 
 
@@ -133,25 +151,21 @@ def property_type(prop):
 def resolve_ref(ref):
     """ Resolve definition reference """
     assert ref.startswith(DEFINITION_BASE_PATH)
-    return TX['definitions'][ref[len(DEFINITION_BASE_PATH):]]
+    return TX_SCHEMA['definitions'][ref[len(DEFINITION_BASE_PATH):]]
 
 
 def main():
     """ Main function """
+    defs = TX_SCHEMA['definitions']
     doc = TPL_DOC % {
-        'wrapper': render_section('transaction', TX),
+        'wrapper': render_section('transaction', TX_SCHEMA),
         'transaction': render_section('transaction',
-                                      TX['properties']['transaction']),
-        'condition': render_section('conditions',
-                                    TX['definitions']['condition']),
-        'fulfillment': render_section('fulfillment',
-                                      TX['definitions']['fulfillment']),
-        'asset': render_section('asset',
-                                TX['definitions']['asset']),
-        'metadata': render_section('metadata',
-                                   TX['definitions']['metadata']['anyOf'][0]),
+                                      TX_SCHEMA['properties']['transaction']),
+        'condition': render_section('conditions', defs['condition']),
+        'fulfillment': render_section('fulfillment', defs['fulfillment']),
+        'asset': render_section('asset', defs['asset']),
+        'metadata': render_section('metadata', defs['metadata']['anyOf'][0]),
         'file': os.path.basename(__file__),
-        'date': datetime.datetime.now().isoformat(),
     }
 
     path = os.path.join(os.path.dirname(__file__), 'source/schema.rst')
