+++ docs/server/source/schema.rst
@@ -1,4 +1,4 @@
-..  This file was auto generated by generate_schema_documentation.py on 2016-11-17T15:57:38.253860
+..  This file was auto generated by generate_schema_documentation.py
 
 ==================
 Transaction Schema
